% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{clean_report}
\alias{clean_report}
\title{Selecciona información útil dentro de tabla validada}
\usage{
clean_report(report, label, id)
}
\arguments{
\item{report}{Data frame}

\item{label}{String. Etiqueta para nombrar a cada una de las columnas}

\item{id}{Identificador de cada registro. Es un string}
}
\description{
Recibe una base validada y cosnerva solo la información útil
}
