% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_results}
\alias{plot_results}
\title{Grafica una columna del data frame}
\usage{
plot_results(data, var)
}
\arguments{
\item{data}{Data frame}

\item{var}{Columna del data frame que se quiere graficar}
}
\description{
Valida una columna del data frame, mostrando la cantidad de casos válidos e inválidos.
}
