% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_functions.R
\name{validate_data}
\alias{validate_data}
\title{Valida un conjunto de columnas de un data frames}
\usage{
validate_data(data, vals, id)
}
\arguments{
\item{data}{Data frame}

\item{vals}{Conjunto de reglas generado con paquete validate. Un objeto por columna}

\item{id}{Es un string. Identificador de cada registro.}
}
\description{
Valida un conjunto de columnas de un data frames, siguiendo las reglas establecidas con anterioridad mediante el paquete validate.
}
