% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_functions.R
\name{validate_var}
\alias{validate_var}
\title{Valida la columna de un data frame, en base a reglas establecidas}
\usage{
validate_var(data, validator, id)
}
\arguments{
\item{data}{Data frame}

\item{validator}{Objeto creado con validate. Conjunto de reglas que debe ser creado previamente para cada una de las columnas.}

\item{id}{Identificador de cada registro. Es un string que debe ser creado dentro de la tabla que se está validando.}
}
\value{
Retorna un \code{data.frame} con información de verdadero o falso para cada uno de los registros
\itemize{
\item id2
\item name
\item value
\item expression
\item all_rules_true
}
}
\description{
Valida una columna de un \code{data.frame}, siguiendo las reglas establecidas con anterioridad.
}
