% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_results}
\alias{plot_results}
\title{\code{plot_results} grafica la cantidad de registros verdaderos y falsos para una columna}
\usage{
plot_results(data, var)
}
\arguments{
\item{data}{\code{data.frame}}

\item{var}{Columna del data frame que se quiere graficar}
}
\value{
Gráfico de \code{ggplot} con registros válidos e inválidos
}
\description{
Muestra un gráfico de \code{ggplot} con la cantidad de registros válidos e inválidos para una columna en un \code{data.frame}.
}
