% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_functions.R
\name{validate_var}
\alias{validate_var}
\title{Valida la columna de un data frame, en base a reglas establecidas}
\usage{
validate_var(data, validator, id)
}
\arguments{
\item{data}{Data frame}

\item{validator}{Conjunto de reglas generado con paquete validate}

\item{id}{Identificador de cada registro. Es un string}
}
\description{
Valida una columna de un data frame, siguiendo las reglas establecidas con anterioridad.
}
